Check if is set idField in methods where is it used.

+++ src/Storage/ALowLevelRelationalDatabaseStorage.php
@@ -75,6 +75,10 @@ abstract class ALowLevelRelationalDatabaseStorage implements ILowLevelRelational
 			throw new \InvalidArgumentException('Argument [id] must be scalar.' . gettype($id) . 'given.');
 		}
 
+		if (is_null($this->idField)) {
+			throw new \InvalidArgumentException("Table primary key is not specified. You cannot use method " . __METHOD__ . ".");
+		}
+
 		$result = $this->connection->select('*')
 			->from($this->tableName)
 			->where($this->idField . ' = ?', $id)
@@ -121,6 +125,10 @@ abstract class ALowLevelRelationalDatabaseStorage implements ILowLevelRelational
 	 */
 	public function find(iterable $ids, ?int $limit = null, ?int $offset = null): IDatabaseDataCollection
 	{
+		if (is_null($this->idField)) {
+			throw new \InvalidArgumentException("Table primary key is not specified. You cannot use method " . __METHOD__ . ".");
+		}
+
 		$query = $this->connection->select('*')
 			->from($this->tableName)
 			->where($this->idField . 'IN(?)', $ids)
@@ -174,6 +182,10 @@ abstract class ALowLevelRelationalDatabaseStorage implements ILowLevelRelational
 	 */
 	public function update($id, iterable $data): int
 	{
+		if (is_null($this->idField)) {
+			throw new \InvalidArgumentException("Table primary key is not specified. You cannot use method " . __METHOD__ . ".");
+		}
+
 		$result = $this->connection->update($this->tableName, $data)
 			->where($this->idField . ' = ?', $id)
 			->execute(\dibi::AFFECTED_ROWS);
@@ -209,6 +221,10 @@ abstract class ALowLevelRelationalDatabaseStorage implements ILowLevelRelational
 	 */
 	public function delete($id): int
 	{
+		if (is_null($this->idField)) {
+			throw new \InvalidArgumentException("Table primary key is not specified. You cannot use method " . __METHOD__ . ".");
+		}
+
 		$result =  $this->connection->delete($this->tableName)
 			->where($this->idField . ' = ?', $id)
 			->execute(\dibi::AFFECTED_ROWS);
